Add basic networking features
Add LRM transport in order to connect to the relay
server, add TextMeshPro for better font handling,
add ParrelSync to be able to open two instances of
the same project for better flexibity when working
on it.

Also add a new Menu Scene that features a basic
lobby creation system.

The player become a prefab to be instanciated and
controlled through the server, but still lacks
proper network controls.

As a result, is possible to create and connect to
lobbies as long as the relay server is up and running.